+++ empmgmnt/src/main/java/com/aks/empmgmnt/entity/Department.java
@@ -14,15 +14,15 @@ public class Department {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	private Integer deptId;
+	private Long deptId;
 
 	private String deptName;
 
-	public Integer getDeptId() {
+	public Long getDeptId() {
 		return deptId;
 	}
 
-	public void setDeptId(Integer deptId) {
+	public void setDeptId(Long deptId) {
 		this.deptId = deptId;
 	}
